requirements.txt
numpy
pandas
tensorflow
tensorflow-addons
matplotlib
opencv-python
pillow
scikit-learn
requests
torch
torchvision
albumentations
tqdm
seaborn
imutils
imageio
pyyaml
scipy
notebook
